사후 체크리스트 요약 및 사후 가이드 추가

+++ src/main/java/dev/jpa/team2/checklist/admin/dto/AdminPostDecisionScoreRowDto.java
@@ -9,5 +9,5 @@ public class AdminPostDecisionScoreRowDto {
 
     private Long itemId;
     private String title;
-    private Double importanceScore;
+    private Integer importanceScore;
 }
