+++ src/main/java/dev/jpa/team2/checklist/ai/dto/ChecklistScoreResult.java
@@ -12,6 +12,6 @@ public class ChecklistScoreResult {
 
     private Long itemId;
     private String title;
-    private Double importanceScore;
+    private Integer importanceScore;
     private String reason;
 }
